Refactoring of selenium
