+++ scripts/selenium/youtube_search.js
@@ -1,14 +1,15 @@
 
-const path = require('path');
 const fs = require('fs');
 const { promisify } = require('util');
+
 const webdriver = require('selenium-webdriver');
-const { Builder, By, until } = webdriver;
-const { ServiceBuilder } = require('selenium-webdriver/chrome');
+const { By, until } = webdriver;
+
+const helper = require('./helper');
 
 
 const youtube_search = async () => {
-  const driver = await create_driver();
+  const driver = await helper.create_driver();
 
   // Refer: https://qiita.com/tonio0720/items/70c13ad304154d95e4bc
   await driver.get('https://www.youtube.com/');
@@ -18,31 +19,9 @@ const youtube_search = async () => {
   let base64 = await driver.takeScreenshot();
   let buffer = Buffer.from(base64, 'base64');
 
-  await promisify(fs.writeFile)('screenshot.jpg', buffer);
+  await promisify(fs.writeFile)('capture.jpg', buffer);
 
   await driver.quit();
-};
-
-
-const create_driver = async () => {
-  const capabilities = webdriver.Capabilities.chrome();
-  capabilities.set('chromeOptions', {
-    args: [
-      '--headless',
-      '--no-sandbox',
-      '--disable-gpu',
-      `--window-size=1980,1200`
-      // other chrome options
-    ]
-  });
-
-  const CHROMEDRIVER_EXE = process.platform === 'win32' ? 'chromedriver.exe' : 'chromedriver';
-  const chromedriver_path = path.join(__dirname, '..', '..', 'bin', CHROMEDRIVER_EXE);
-  const service_builder = new ServiceBuilder(chromedriver_path);
-  const builder = new Builder().withCapabilities(capabilities).setChromeService(service_builder);
-  const driver = await builder.build();
-
-  return driver;
 };
